requirements.txt
python-vlc
gpiozero